+++ Assets/Scenes/FinalScene.unity
@@ -875,6 +875,28 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: c65d93c08c5dd6e49af614e6e61abf1e, type: 3}
+--- !u!1 &935294075 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 544493633091346230, guid: f50bf5bc061c3434ead33d97c799ee9f,
+    type: 3}
+  m_PrefabInstance: {fileID: 5528007982197585781}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &935294077
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 935294075}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d264b1476446b4bc1bff894fa1fae2d9, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  wayPoints: []
+  velocity: 3
+  waypointNumber: 0
+  fieldOfView: {fileID: 0}
 --- !u!1001 &1081012905
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -1383,7 +1405,11 @@ PrefabInstance:
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 544493633091346230, guid: f50bf5bc061c3434ead33d97c799ee9f,
+        type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 935294077}
   m_SourcePrefab: {fileID: 100100000, guid: f50bf5bc061c3434ead33d97c799ee9f, type: 3}
 --- !u!1001 &5606351299714451752
 PrefabInstance:
