+++ Assets/Scripts/States/Leader.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Enemy : MonoBehaviour
+public class Leader : MonoBehaviour
 {
     private FiniteStateMachine fsm;
     [SerializeField] protected List<Node> wayPoints;
@@ -20,11 +20,11 @@ public class Enemy : MonoBehaviour
 
         fsm.AddState(EnemyState.BackToPatrol, new EnemyBackToPatrol());
         fsm.AddState(EnemyState.Follow, new EnemyFollow());
-        fsm.AddState(EnemyState.Patrol, new EnemyPatrolState());
+        fsm.AddState(EnemyState.Patrol, new LeaderIdleState());
         fsm.ChangeState(EnemyState.Patrol, transform.position);
         EnemyFollow.onFoundPlayer += SetFollowState;
-
     }
+
     public void SetFollowState(Vector3 target)
     {
         if (!(fsm.GetCurrentState() is EnemyFollow))
@@ -33,14 +33,28 @@ public class Enemy : MonoBehaviour
         }
     }
 
-
-    // Update is called once per frame
     void Update()
     {
         fsm.Update();
+
+        if (Input.GetMouseButtonDown(0))
+        {
+            Vector3 click = Camera.main.ScreenToWorldPoint(Input.mousePosition);
+            click = new Vector3(click.x, click.y, 0);
+
+            fsm.ChangeState(EnemyState.Chase, click);
+        }
+
+        if (Input.GetMouseButtonDown(1))
+        {
+
+        }
+        if (Input.GetMouseButtonDown(2))
+        {
+            
+        }
     }
 
-    
     public List<Node> GetWayPoints()
     {
         return this.wayPoints;
@@ -51,11 +65,12 @@ public class Enemy : MonoBehaviour
         transform.forward= dir;
         transform.position += dir.normalized * velocity * Time.deltaTime;
     }
-
+    
     public GameObject GetTargetPlayer()
     {
         return fieldOfView.FieldOfView();
     }
+
     public int GetWayPointNumber()
     {
         return waypointNumber;
